qutr-3-assignments/Custom-Web-Search-Tool/my-agent/requirements.txt
openai>=1.40.0
guardrails-ai>=0.5.0
requests
python-dotenv
